% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceable.R
\name{attach_source_dep}
\alias{attach_source_dep}
\title{Attach dependencies to source code
Attach dependencies to the source code (any input variables are automatically
attached)}
\usage{
attach_source_dep(x, deps)
}
\description{
Attach dependencies to source code
Attach dependencies to the source code (any input variables are automatically
attached)
}
\keyword{internal}
