% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkm.R
\docType{data}
\name{GeomKm}
\alias{GeomKm}
\alias{GeomKmband}
\alias{GeomKmticks}
\alias{StatKm}
\alias{StatKmband}
\alias{ggquickeda-ggproto}
\alias{StatKmticks}
\title{Geom Proto}
\description{
Geom Proto

Geom Proto

Geom Proto

Geom Proto

Geom Proto

Geom Proto
}
\keyword{internal}
