% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceable.R
\name{+}
\alias{+}
\title{Add 'ggplot2' layer to a sourceable 'ggplot2'}
\usage{
`+`(e1, e2)
}
\description{
Overwrite the plus operator so that if a "sourceable" object is used,
the source code is kept
}
\keyword{internal}
