% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ggquickeda.R
\name{run_ggquickeda}
\alias{run_ggquickeda}
\title{Run the ggquickeda application}
\usage{
run_ggquickeda(data = NULL, ...)
}
\arguments{
\item{data}{The initial data.frame to load into the application.}

\item{...}{Additional arguments for bookmarking}
}
\description{
Run the \code{ggquickeda} application.
}
\examples{
if (interactive()) {
  run_ggquickeda()
}
}
