% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmrisktable.R
\name{ggkmrisktable}
\alias{ggkmrisktable}
\title{ggkmrisktable}
\usage{
ggkmrisktable(
  data = lung_long,
  time = "time",
  status = "DV",
  endpoint = "Endpoint",
  groupvar1 = "Endpoint",
  groupvar2 = "expname",
  exposure_metrics = c("age", "ph.karno"),
  exposure_metric_split = c("median", "tertile", "quartile", "none"),
  exposure_metric_soc_value = -99,
  exposure_metric_plac_value = 0,
  color_fill = "exptile",
  linetype = "exptile",
  xlab = "Time of follow_up",
  ylab = "Overall survival probability",
  nrisk_table_plot = TRUE,
  nrisk_table_variables = c("n.risk"),
  nrisk_table_breaktimeby = NULL,
  nrisk_table_textsize = 4,
  nrisk_position_scaler = 0.2,
  nrisk_position_dodge = 0.2,
  nrisk_offset = 0,
  nrisk_filterout0 = FALSE,
  km_logrank_pvalue = FALSE,
  km_trans = "identity",
  km_ticks = TRUE,
  km_band = TRUE,
  km_conf_int = 0.95,
  km_conf_type = "log",
  km_conf_lower = "usual",
  km_median = "none",
  km_median_pos = "left",
  km_yaxis_position = c("left"),
  facet_formula = NULL,
  facet_ncol = NULL,
  facet_strip_position = c("top", "top", "top", "top"),
  theme_certara = TRUE
)
}
\arguments{
\item{data}{Data to use with multiple endpoints stacked}

\item{time}{name of the column holding the time to event information default to `time`}

\item{status}{name of the column holding the event information default to `DV`}

\item{endpoint}{name of the column holding the name/key of the endpoint default to `Endpoint`}

\item{groupvar1}{name of the column to group by default `Endpoint`}

\item{groupvar2}{name of the column to group by in addition to endpoint and `expname`}

\item{exposure_metrics}{name(s) of the column(s) to be stacked into `expname` `exptile` and split into `exposure_metric_split`}

\item{exposure_metric_split}{Possible values: "median","tertile","quartile","none"}

\item{exposure_metric_soc_value}{special exposure code for standard of care default -99}

\item{exposure_metric_plac_value}{special exposure code for placebo default 0}

\item{color_fill}{name of the column to be used for color/fill default to `exptile`}

\item{linetype}{name of the column to be used for linetype default to `exptile`}

\item{xlab}{text to be used as x axis label}

\item{ylab}{text to be used as y axis label}

\item{nrisk_table_plot}{TRUE}

\item{nrisk_table_variables}{c("n.risk"),#n.risk pct.risk n.event cum.n.event n.censor}

\item{nrisk_table_breaktimeby}{NULL}

\item{nrisk_table_textsize}{4}

\item{nrisk_position_scaler}{0.2}

\item{nrisk_position_dodge}{0.2}

\item{nrisk_offset}{0}

\item{nrisk_filterout0}{FALSE}

\item{km_logrank_pvalue}{FALSE}

\item{km_trans}{#"identity","event","cumhaz","cloglog")}

\item{km_ticks}{TRUE}

\item{km_band}{TRUE}

\item{km_conf_int}{0.95}

\item{km_conf_type}{"log", #c("none" , "plain", "log" ,"log-log","logit"),}

\item{km_conf_lower}{"usual", #c("peto" , "modified", "usual"),}

\item{km_median}{add median survival information "none", "median", "medianci", "table"}

\item{km_median_pos}{when table is chosen where to put it}

\item{km_yaxis_position}{where to put y axis on "left" or "right}

\item{facet_formula}{facet formula}

\item{facet_ncol}{NULL}

\item{facet_strip_position}{c("top","top","top","top")}

\item{theme_certara}{apply certara colors and format for strips and default colour/fill}
}
\description{
Produces a km plot with a facettable risk table in ggplot2
}
\examples{
library(tidyr)
# Example 1
lung_long <-  survival::lung |>
 dplyr::mutate(status = ifelse(status==1,0,1)) |>
 tidyr::gather(Endpoint,DV,status) |>
 dplyr::filter(!is.na(ph.karno))|>
 dplyr::filter(!is.na(pat.karno))|>
 dplyr::filter(!is.na(ph.ecog))
lung_long$ph.ecog <- ifelse(lung_long$ph.ecog>1,2,lung_long$ph.ecog)
lung_long$ph.ecog <- as.factor(lung_long$ph.ecog )
lung_long$ph.ecog <- as.factor(lung_long$ph.ecog )
lung_long$facetdum <- "(all)"

ggkmrisktable(data = lung_long, time= "time", status ="DV",
             exposure_metrics =c("age","ph.karno"),
             exposure_metric_split = "tertile",
             color_fill = "exptile",
             linetype = "expname",
             groupvar1 = "Endpoint",
             groupvar2 = "exptile",
             xlab = "Time of follow_up",
             ylab ="Overall survival probability",
             nrisk_table_variables = c("n.risk","n.event"),
             km_median = "medianci",
             km_band = FALSE,
             nrisk_table_breaktimeby = 200,
             facet_ncol = 3)
#Example 2
ggkmrisktable(data = lung_long, time= "time", status ="DV",
             exposure_metrics =c("age","ph.karno"),
             exposure_metric_split = "quartile",
             color_fill = "exptile",
             linetype = "none",
             groupvar1 = "Endpoint",
             groupvar2 = "exptile",
             xlab = "Time of follow_up",
             ylab ="Overall survival probability",
             nrisk_table_variables = c("cum.n.event","pct.risk","n.censor"),
             km_median = "medianci",
             km_band = TRUE,
             km_trans = "event",
             nrisk_table_breaktimeby = 200,
             facet_ncol = 3,
             facet_formula = ~expname)
#Example 3
ggkmrisktable(data = lung_long, time = "time", status = "DV",
             exposure_metrics =c("ph.karno","pat.karno"),
             exposure_metric_split = "median",
             color_fill = "exptile",
             linetype = "none",
             groupvar1 = "Endpoint",
             groupvar2 = "expname",
             xlab = "Time of follow_up",
             ylab ="Overall survival probability",
             nrisk_table_variables = c("n.event"),
             km_trans = "event",
             km_median = "table",
             km_median_pos = "right",
             km_logrank_pvalue = TRUE,
             km_band = TRUE,
             nrisk_table_breaktimeby = 200,
             facet_ncol = 3,
             facet_formula = ~expname)
#Example 4
ggkmrisktable(data=lung_long,
             exposure_metrics = c("ph.karno","age"),
             exposure_metric_split = "median",
             time = "time",
             status ="DV",
             color_fill = "ph.ecog",
             linetype = "ph.ecog",
             groupvar1 = "exptile",
             groupvar2 = "expname",
             nrisk_filterout0 = FALSE,
             nrisk_table_breaktimeby = 200,
             km_logrank_pvalue = TRUE,
             km_median = "table",
             km_median_pos = "left",
             facet_formula = ~expname+exptile)
}
